\name{trajectorySim-package}
\alias{trajectorySim-package}
\alias{trajectorySim}
\docType{package}
\title{
Trajectory Similarity
}
\description{
Similarity measures for trajectories, including Hausdorff, equal-time, Fréchet distances, dynamic time warping (DTW), longest common subsequence (LCSS), spatiotemporal locality inside polylines (STLIP).
Arbitrary dimension, flexible distance metric.
}
\details{
\tabular{ll}{
Package: \tab trajectory-sim\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-15\cr
License: \tab GPL (>= 3)\cr
}
#~~ An overview of how to use the package, including the most important ~~
#~~ functions ~~
}
\author{
Dongliang Peng <dongliang.peng@uni-wuerzburg.de>, 
Stef Sijben <Stef.Sijben@ruhr-uni-bochum.de>
Maintainer: Stef Sijben <Stef.Sijben@ruhr-uni-bochum.de>
}
\references{
}
\keyword{ package }
\seealso{
\code{\link[move:move-package]{move}}
}
\examples{

}
